Add Update and Delete authentication

+++ app/Controllers/Http/ContactsController.ts
@@ -1,3 +1,4 @@
+import { Exception } from '@poppinss/utils'
 import Contact from 'App/Models/Contact'
 import Todo from 'App/Models/Todo'
 
@@ -27,11 +28,21 @@ export default class ContactsController {
 
   public async destroy(userId: number, id: number) {
     const contact = await Contact.findByOrFail('id', id)
-    return await contact.delete()
+
+    if (userId === contact.userId) { 
+      return await contact.delete()
+    } else {
+      throw new Exception('Not allowed')
+    }
   }
 
   public async update(userId: number, data: Record<string, any>) {
     const contact = await Contact.findByOrFail('id', data.id)
+
+    if (!!userId) {
+      await contact.associateUser(userId)
+      await contact.load('user')
+    }
 
     return await contact.merge(data).save()
   }
